correct positions when text size the same, 10
