+++ src/components/editor/jspdf.js
@@ -1,56 +1,96 @@
-// import React, { PureComponent } from 'react';
-// import { Button, } from '@material-ui/core';
+import React, { PureComponent } from 'react';
+import { Button, } from '@material-ui/core';
 
-// import jsPDF from 'jspdf'
-// let data = {
-//     "time": 1554920381017,
-//     "blocks": [
-//         {
-//             "type": "header",
-//             "data": {
-//                 "text": "Hello Editor.js",
-//                 "level": 2
-//             }
-//         },
-//         {
-//             type: "paragraph",
-//             data: {
-//                 text:
-//                     "We have been working on this project more than three years. Several large media projects help us to test and debug the Editor, to make it's core more stable. At the same time we significantly improved the API. Now, it can be used to create any plugin for any task. Hope you enjoy."
-//             }
-//         },
-//     ],
-//     "version": "2.12.4"
-// }
-// jsPdfGenerator = () => {
+import jsPDF from 'jspdf'
+let data = {
+    "time": 1554920381017,
+    "blocks": [
 
-//     var doc = new jsPDF('p', 'mm', [297, 210]); // creates A4 format page
+        {
+            type: "paragraph",
+            data: {
+                text:
+                    "This is a test sentence number one"
+            }
+        },
+        {
+            type: "paragraph",
+            data: {
+                text:
+                    "This is a test sentence number two"
+            }
+        },
 
-//     doc.setFont('Roboto-Regular');
-//     doc.setTextColor(255, 0, 0)
-//     // doc.setCharSpace(0.3)
-//     // doc.setFontStyle("italic")
-//     // doc.autoPrint()
-//     // doc.setFontSize(30)
-//     // doc.setFont('bold');
-//     data.blocks.forEach((block, i) => {
-//         block.type === "header" ? doc.setFontSize(30) : doc.setFontSize(10)
-//         var splitText = doc.splitTextToSize(block.data.text, 180);
-//         doc.text(10, 10 + (i * 10), splitText)
-//     })
-//     doc.save('editor.pdf');
-// }
-
-
-// function pdfGenerate() {
-
-//     // JSpdf Generator For generating the PDF
+        {
+            type: "paragraph",
+            data: {
+                text:
+                    "This is a test sentence number three"
+            }
+        },
+        {
+            type: "paragraph",
+            data: {
+                text:
+                    "This is a test sentence number four"
+            }
+        },
+        {
+            type: "paragraph",
+            data: {
+                text:
+                    "This is a test sentence number five"
+            }
+        }
 
 
-//     return (
-//         <Button onClick={jsPdfGenerator} > Generate PDF </Button>
-//     )
+    ],
+    "version": "2.12.4"
+}
+
+export default class pdfGenerate extends PureComponent {
+
+    // JSpdf Generator For generating the PDF
+    jsPdfGenerator = () => {
+        let numOfLines = 0;
+
+        let next, tmp, cur = 10;
+        next = cur;
+
+        var doc = new jsPDF('p', 'mm', [297, 210]); // creates A4 format page
+        // 
+        doc.setFont('Roboto-Regular');
+        // doc.setTextColor(255, 0, 0)
+
+        data.blocks.forEach((block, i) => {
+            numOfLines = 0;
+
+            doc.setFontSize(10)
+            var splitText = doc.splitTextToSize(block.data.text, 180);
+            console.log('split text length', splitText.length)
+            numOfLines = splitText.length
+
+            // write on next availible pos
+            console.log('now printing at post Y:', cur)
+            doc.text(10, cur, splitText)
+            // recalculate next 
 
 
-// }
-// export default pdfGenerate;
+            next = next + (5 * numOfLines)
+            console.log('next block will be at Y:', next)
+            // recalculate previous
+
+            cur = next;
+            console.log('previous pos Y:', cur)
+
+        })
+        doc.save('editor.pdf');
+    }
+
+    render() {
+        return (
+            <Button onClick={this.jsPdfGenerator}  > Generate PDF </Button>
+        )
+    }
+
+}
